add map in sub-component #38
